refactor: progress change to ts
