+++ components/progress/index.tsx
@@ -1,9 +1,10 @@
+import { App } from 'vue';
 import Progress from './progress';
 
 export { ProgressProps } from './progress';
 
 /* istanbul ignore next */
-Progress.install = function(app) {
+Progress.install = function(app: App) {
   app.component(Progress.name, Progress);
 };
